model/vision-model/requirements.txt
ultralytics
opencv-python
websockets
tensorflow<2.11
keras
numpy
Pillow
h5py